+++ EasyPost/BetaClient.cs
@@ -1,6 +1,6 @@
 using System.Net.Http;
 using EasyPost._base;
-using EasyPost.Services.Beta;
+using EasyPost.Services;
 
 namespace EasyPost
 {
@@ -21,8 +21,6 @@ namespace EasyPost
          *
          * When you are migrating a service from beta to general, you should remove/change the overrideApiVersion parameter from each function that is being migrated.
          */
-
-        public EndShipperService EndShipper => GetService<EndShipperService>();
 
         public PartnerService Partner => GetService<PartnerService>();